HXENG-64 refactor ATS prior to fork
- Pick up the t-core model classes that have moved or been renamed
