+++ pom.xml
@@ -52,7 +52,7 @@
         <dependency.activiti-engine.version>5.23.0</dependency.activiti-engine.version>
         <dependency.activiti.version>5.23.0</dependency.activiti.version>
         <dependency.alfresco-transform-service.version>1.5.3</dependency.alfresco-transform-service.version>
-        <dependency.alfresco-transform-core.version>2.6.0</dependency.alfresco-transform-core.version>
+        <dependency.alfresco-transform-core.version>3.0.0-HXP-A4</dependency.alfresco-transform-core.version>
         <dependency.alfresco-greenmail.version>6.4</dependency.alfresco-greenmail.version>
         <dependency.acs-event-model.version>0.0.16</dependency.acs-event-model.version>
