+++ repository/src/main/java/org/alfresco/repo/content/transform/LocalTransformImpl.java
@@ -31,7 +31,7 @@ import org.alfresco.service.cmr.repository.ContentReader;
 import org.alfresco.service.cmr.repository.ContentWriter;
 import org.alfresco.service.cmr.repository.MimetypeService;
 import org.alfresco.service.cmr.repository.NodeRef;
-import org.alfresco.transform.client.model.config.TransformOption;
+import org.alfresco.transform.config.TransformOption;
 import org.alfresco.util.Pair;
 
 import java.util.HashMap;
