+++ repository/src/main/java/org/alfresco/repo/content/transform/LocalTransformServiceRegistry.java
@@ -36,14 +36,14 @@ import java.util.function.Consumer;
 import java.util.stream.Collectors;
 
 import org.alfresco.service.cmr.repository.MimetypeService;
-import org.alfresco.transform.client.model.config.CoreFunction;
-import org.alfresco.transform.client.model.config.TransformOptionGroup;
-import org.alfresco.transform.client.registry.CombinedConfig;
-import org.alfresco.transform.client.model.config.TransformOption;
-import org.alfresco.transform.client.registry.TransformServiceRegistryImpl;
-import org.alfresco.transform.client.model.config.TransformStep;
-import org.alfresco.transform.client.model.config.Transformer;
-import org.alfresco.transform.client.registry.TransformServiceRegistry;
+import org.alfresco.transform.config.CoreFunction;
+import org.alfresco.transform.config.TransformOptionGroup;
+import org.alfresco.transform.registry.CombinedConfig;
+import org.alfresco.transform.config.TransformOption;
+import org.alfresco.transform.registry.TransformServiceRegistryImpl;
+import org.alfresco.transform.config.TransformStep;
+import org.alfresco.transform.config.Transformer;
+import org.alfresco.transform.registry.TransformServiceRegistry;
 import org.alfresco.util.PropertyCheck;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
