+++ repository/src/main/java/org/alfresco/repo/rendition2/SwitchingTransformServiceRegistry.java
@@ -25,8 +25,8 @@
  */
 package org.alfresco.repo.rendition2;
 
-import org.alfresco.transform.client.model.config.CoreFunction;
-import org.alfresco.transform.client.registry.TransformServiceRegistry;
+import org.alfresco.transform.config.CoreFunction;
+import org.alfresco.transform.registry.TransformServiceRegistry;
 
 import java.util.Map;
