+++ repository/src/test/java/org/alfresco/AppContext06TestSuite.java
@@ -25,6 +25,7 @@
  */
 package org.alfresco;
 
+import org.alfresco.transform.registry.LocalTransformServiceRegistryConfigTest;
 import org.alfresco.util.testing.category.DBTests;
 import org.alfresco.util.testing.category.NonBuildTests;
 import org.junit.experimental.categories.Categories;
@@ -50,7 +51,7 @@ import org.junit.runners.Suite;
     org.alfresco.repo.event2.RepoEvent2ITSuite.class,
 
     // Requires running transformers
-    org.alfresco.transform.client.registry.LocalTransformServiceRegistryConfigTest.class,
+    LocalTransformServiceRegistryConfigTest.class,
     org.alfresco.repo.rendition2.RenditionService2IntegrationTest.class,
     org.alfresco.repo.rendition2.LocalTransformServiceRegistryIntegrationTest.class,
     org.alfresco.repo.rendition2.LocalTransformClientIntegrationTest.class,
