+++ repository/src/test/java/org/alfresco/repo/rendition2/RenditionService2Test.java
@@ -40,7 +40,7 @@ import org.alfresco.service.cmr.repository.NodeService;
 import org.alfresco.service.cmr.rule.RuleService;
 import org.alfresco.service.namespace.QName;
 import org.alfresco.service.transaction.TransactionService;
-import org.alfresco.transform.client.registry.TransformServiceRegistryImpl;
+import org.alfresco.transform.registry.TransformServiceRegistryImpl;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
